+++ src/main/java/be/swsb/fiazard/ordering/orderplacement/OrderFactory.java
@@ -1,4 +1,4 @@
-package be.swsb.fiazard.bestelling;
+package be.swsb.fiazard.ordering.orderplacement;
 
 import com.google.common.base.Preconditions;
