+++ src/main/java/be/swsb/fiazard/ordering/orderplacement/OrderingBundle.java
@@ -1,4 +1,4 @@
-package be.swsb.fiazard.bestelling;
+package be.swsb.fiazard.ordering.orderplacement;
 
 import be.swsb.fiazard.ddd.AggregateIdGenerator;
 import be.swsb.fiazard.eventstore.AggregateRepository;
